move/copy some more tests
